extract trains array from UI to Interactor
